requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
openai
numpy
python-dotenv
requests
